requirements.txt
Keras==2.2.0
Keras-Applications==1.0.2
Keras-Preprocessing==1.0.1
matplotlib==2.2.2
numpy==1.13.3
pandas==0.22.0
Pillow==4.3.0
pydicom==1.1.0
scikit-image==0.14.1
scikit-learn==0.19.1
scipy==1.0.1
